auto_return and import fix
